requirements.txt
Pillow==9.5.0
qrcode==7.4.2
CairoSVG==2.7.1